+++ src/dev_timer.h
@@ -8,17 +8,17 @@
 namespace DiaDevice {
     class Timer {
     private:
-        long            ms; // Milliseconds
-        time_t          s;  // Seconds
-        struct timespec spec;
-        long            coef;
+        long long           ms; // Milliseconds
+        time_t              s;  // Seconds
+        struct timespec     spec;
+        long long           coef;
         
     public:
-        long CurrentTimeMS() {
+        long long CurrentTimeMS() {
             clock_gettime(CLOCK_REALTIME, &spec);
             s  = spec.tv_sec;
-            ms = (long)spec.tv_nsec / ONE_MILLION; // Convert nanoseconds to milliseconds
-            return s*(long)1000 + ms;
+            ms = (long long)spec.tv_nsec / ONE_MILLION; // Convert nanoseconds to milliseconds
+            return s*(long long)1000 + ms;
         }
     };
 }
